requirements.txt
annotated-types==0.7.0
anyio==4.5.2
beautifulsoup4==4.13.3
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
duckdb==1.2.0
exceptiongroup==1.2.2
Faker==35.2.2
fastapi==0.115.8
flask==3.0.3
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
joblib==1.4.2
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
PyAudio==0.2.14
pydantic==2.10.6
pydantic-core==2.27.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.10.4
starlette==0.44.0
threadpoolctl==3.5.0
typing==3.7.4.3
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.33.0
werkzeug==3.0.6
zipp==3.20.2
